pyproject.toml
[tool.poetry]
name = "pytoncenter"
version = "0.0.5"
description = ""
authors = ["alan890104 <alan890104@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
tonpy = "^0.0.0.1.2b0"
aiohttp = "^3.9.1"
pynacl = "^1.5.0"
treelib = "^1.7.0"
pydantic = "^2.6.1"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
datamodel-code-generator = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."pytoncenter/__init__.py"]